requirements.txt
Flask==3.1.0
librosa==0.10.2.post1
openai_whisper==20240930
Requests==2.32.3
soundfile==0.12.1
spacy==3.8.2
